renamed factfinder to sloyka
