+++ sloyka/src/exсeptions_city.csv
